LPS-198311 refactor MultiVMEhcachePortalCacheManagerConfigurator.java
move class to the only usage, and delete the other class
